Re-enable ADAM for 1D (non-angular) analysis (#42)
* refactoring and adding adam

* fitting epw

* previous weights

* angular

* angular

* angular

* angular

* num epochs

* plot

* sigma

* no static

* list dict

* list dict

* deepcopy

* deepcopy

* deepcopy

* deepcopy

* angular running without error bars

* smooth

* removing some code

* dist and param handling

* shapes..

* jit

* num mins

* epw test
