+++ configs/1d/inputs.yaml
@@ -86,28 +86,29 @@ parameters:
         symmetric: False
 
 data:
-    shotnum: 106315
+    shotnum: 101675
     lineouts:
         type:
-            um
-        start: -1000
-        end: 1000
-        skip: 20
+            pixel
+        start: 340
+        end: 660
+        skip: 4
     background:
         type:
             Fit
         slice: 900
+        show: False
 
 other:
     extraoptions:
         load_ion_spec: True
         load_ele_spec: True
-        fit_IAW: True
-        fit_EPWb: False
-        fit_EPWr: False
+        fit_IAW: False
+        fit_EPWb: True
+        fit_EPWr: True
     refit: False
     refit_thresh: 5.0
-    calc_sigmas: True
+    calc_sigmas: False
 
 
 mlflow:
