+++ inverse_thomson_scattering/fitter.py
@@ -1,11 +1,13 @@
+import copy
 from typing import Dict, Tuple
 import time
 import numpy as np
 import pandas as pd
 import scipy.optimize as spopt
 
-import jaxopt, mlflow  # , optax
+import jaxopt, mlflow, optax
 from tqdm import trange
+from jax.flatten_util import ravel_pytree
 
 from inverse_thomson_scattering.misc.num_dist_func import get_num_dist_func
 from inverse_thomson_scattering.model.loss_function import TSFitter
@@ -13,7 +15,17 @@ from inverse_thomson_scattering.process import prepare, postprocess
 
 
 def init_param_norm_and_shift(config: Dict) -> Dict:
-    # init_params = {}
+    """
+    Initializes the dictionary that contains the normalization constants for all the parameters
+
+    The parameters are all normalized from 0 to 1 in order to improve gradient flow
+
+    Args:
+        config: Dict
+
+    Returns: Dict
+
+    """
     lb = {}
     ub = {}
     parameters = config["parameters"]
@@ -41,7 +53,16 @@ def init_param_norm_and_shift(config: Dict) -> Dict:
     return {"norms": norms, "shifts": shifts, "lb": lb, "ub": ub}
 
 
-def validate_inputs(config):
+def _validate_inputs_(config: Dict) -> Dict:
+    """
+    This function adds derived configuration quantities that are necessary for the fitting process
+
+    Args:
+        config: Dict
+
+    Returns: Dict
+
+    """
     # get derived quantities
     config["velocity"] = np.linspace(-7, 7, config["parameters"]["fe"]["length"])
     if config["parameters"]["fe"]["symmetric"]:
@@ -81,7 +102,21 @@ def validate_inputs(config):
     return config
 
 
-def scipy_angular_loop(config, all_data, best_weights, all_weights, sa):
+def scipy_angular_loop(config: Dict, all_data: Dict, sa):
+    """
+    Performs angular thomson scattering i.e. ARTEMIS fitting exercise using the SciPy optimizer routines
+
+
+    Args:
+        config:
+        all_data:
+        best_weights:
+        all_weights:
+        sa:
+
+    Returns:
+
+    """
     print("Running Angular, setting batch_size to 1")
     config["optimizer"]["batch_size"] = 1
     batch = {
@@ -96,21 +131,31 @@ def scipy_angular_loop(config, all_data, best_weights, all_weights, sa):
             config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :
         ],
     }
-    for i in range(config["optimizer"]["num_mins"]):
+
+    ts_fitter = TSFitter(config, sa, batch)
+    all_weights = {k: [] for k in ts_fitter.pytree_weights["active"].keys()}
+    if config["optimizer"]["num_mins"] > 1:
+        print(Warning("multiple num mins doesnt work. only running once"))
+    for i in range(1):
         ts_fitter = TSFitter(config, sa, batch)
+        init_weights = copy.deepcopy(ts_fitter.flattened_weights)
+
         # ts_fitter.flattened_weights = ts_fitter.flattened_weights * np.random.uniform(
         #     0.97, 1.03, len(ts_fitter.flattened_weights)
         # )
         res = spopt.minimize(
             ts_fitter.vg_loss if config["optimizer"]["grad_method"] == "AD" else ts_fitter.loss,
-            ts_fitter.flattened_weights,
+            init_weights,
             args=batch,
             method=config["optimizer"]["method"],
             jac=True if config["optimizer"]["grad_method"] == "AD" else False,
             bounds=ts_fitter.bounds,
-            options={"disp": True, "maxiter": 1},
+            options={"disp": True, "maxiter": config["optimizer"]["num_epochs"]},
         )
-        best_weights = ts_fitter.unravel_pytree(res["x"])
+        these_weights = ts_fitter.unravel_pytree(res["x"])
+        for k in all_weights.keys():
+            all_weights[k].append(these_weights[k])
+
         if i == config["optimizer"]["num_mins"] - 1:
             break
         config["parameters"]["fe"]["length"] = (
@@ -120,12 +165,12 @@ def scipy_angular_loop(config, all_data, best_weights, all_weights, sa):
         if config["parameters"]["fe"]["symmetric"]:
             refined_v = np.linspace(0, 7, config["parameters"]["fe"]["length"])
 
-        refined_fe = np.interp(refined_v, config["velocity"], np.squeeze(best_weights["fe"]))
+        refined_fe = np.interp(refined_v, config["velocity"], np.squeeze(these_weights["fe"]))
 
         config["parameters"]["fe"]["val"] = refined_fe.reshape((1, -1))
         config["velocity"] = refined_v
-        config["parameters"]["ne"]["val"] = best_weights["ne"].squeeze()
-        config["parameters"]["Te"]["val"] = best_weights["Te"].squeeze()
+        config["parameters"]["ne"]["val"] = these_weights["ne"].squeeze()
+        config["parameters"]["Te"]["val"] = these_weights["Te"].squeeze()
 
         config["parameters"]["fe"]["ub"] = -0.5
         config["parameters"]["fe"]["lb"] = -50
@@ -136,19 +181,149 @@ def scipy_angular_loop(config, all_data, best_weights, all_weights, sa):
             config["parameters"]["fe"]["ub"], np.ones(config["parameters"]["fe"]["length"])
         )
         config["units"] = init_param_norm_and_shift(config)
-        all_weights.append(best_weights)
-    raw_weights = ts_fitter.unravel_pytree(res["x"])
+
     overall_loss = res["fun"]
-    best_weights = all_weights
-    # print(best_weights)
 
-    return raw_weights, best_weights, overall_loss, ts_fitter
+    return all_weights, overall_loss, ts_fitter
 
 
-def scipy_1d_loop(config, all_data, batch_indices, num_batches, best_weights, sa):
+def angular_adam(config, all_data, sa, batch_indices, num_batches):
+    """
+    This performs an SGD-like fitting routine. It is different than the SciPy routines primarily because we have to
+    manage the number of iterations manually whereas the SciPy routines have their own convergence/termination process
+
+    Args:
+        config:
+        all_data:
+        sa:
+        batch_indices:
+        num_batches:
+
+    Returns:
+
+    """
+
+    config["optimizer"]["batch_size"] = 1
+    test_batch = {
+        "e_data": all_data["e_data"][config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :],
+        "e_amps": all_data["e_amps"][config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :],
+        "i_data": all_data["i_data"],
+        "i_amps": all_data["i_amps"],
+        "noise_e": config["other"]["PhysParams"]["noiseE"][
+            config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :
+        ],
+        "noise_i": config["other"]["PhysParams"]["noiseI"][
+            config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :
+        ],
+    }
+    func_dict = get_loss_function(config, sa, test_batch)
+    jaxopt_kwargs = dict(
+        fun=func_dict["vg_func"], maxiter=config["optimizer"]["num_epochs"], value_and_grad=True, has_aux=True
+    )
+    opt = optax.adam(config["optimizer"]["learning_rate"])
+    solver = jaxopt.OptaxSolver(opt=opt, **jaxopt_kwargs)
+
+    weights = func_dict["init_weights"]
+    opt_state = solver.init_state(weights, batch=test_batch)
+
+    # start train loop
+    t1 = time.time()
+    print("minimizing")
+    mlflow.set_tag("status", "minimizing")
+
+    best_loss = 1e16
+    for i_epoch in (pbar := trange(config["optimizer"]["num_epochs"])):
+        if config["nn"]["use"]:
+            np.random.shuffle(batch_indices)
+            # tbatch.set_description(f"Epoch {i_epoch + 1}, Prev Epoch Loss {epoch_loss:.2e}")
+        epoch_loss = 0.0
+        weights, opt_state = solver.update(params=weights, state=opt_state, batch=test_batch)
+        epoch_loss += opt_state.value
+        pbar.set_description(f"Loss {epoch_loss:.2e}")
+
+        # epoch_loss /= num_batches
+        if epoch_loss < best_loss:
+            best_loss = epoch_loss
+            best_weights = weights
+
+        mlflow.log_metrics({"epoch loss": float(epoch_loss)}, step=i_epoch)
+
+
+def _1d_adam_loop_(config, ts_fitter, previous_weights, batch, tbatch):
+    jaxopt_kwargs = dict(
+        fun=ts_fitter.vg_loss, maxiter=config["optimizer"]["num_epochs"], value_and_grad=True, has_aux=True
+    )
+    opt = optax.adam(config["optimizer"]["learning_rate"])
+    solver = jaxopt.OptaxSolver(opt=opt, **jaxopt_kwargs)
+
+    if previous_weights is None:
+        init_weights = ts_fitter.pytree_weights["active"]
+    else:
+        init_weights = previous_weights
+
+    # if "sequential" in config["optimizer"]:
+    #     if config["optimizer"]["sequential"]:
+    #         if previous_weights is not None:
+    #             init_weights = previous_weights
+
+    opt_state = solver.init_state(init_weights, batch=batch)
+
+    best_loss = 1e16
+    epoch_loss = 1e19
+    for i_epoch in range(config["optimizer"]["num_epochs"]):
+        tbatch.set_description(f"Epoch {i_epoch + 1}, Prev Epoch Loss {epoch_loss:.2e}")
+        # if config["nn"]["use"]:
+        #     np.random.shuffle(batch_indices)
+
+        init_weights, opt_state = solver.update(params=init_weights, state=opt_state, batch=batch)
+        epoch_loss = opt_state.value
+        if epoch_loss < best_loss:
+            best_loss = epoch_loss
+            best_weights = init_weights
+
+    return best_loss, best_weights
+
+
+def _1d_scipy_loop_(config, ts_fitter: TSFitter, previous_weights, batch, tbatch):
+    if previous_weights is None:  # if prev, then use that, if not then use flattened weights
+        init_weights = np.copy(ts_fitter.flattened_weights)
+    else:
+        init_weights = np.array(previous_weights)
+
+    # if "sequential" in config["optimizer"]:
+    #     if config["optimizer"]["sequential"]:
+    #         if previous_weights is not None:
+    #             init_weights = previous_weights
+
+    res = spopt.minimize(
+        ts_fitter.vg_loss if config["optimizer"]["grad_method"] == "AD" else ts_fitter.loss,
+        init_weights,
+        args=batch,
+        method=config["optimizer"]["method"],
+        jac=True if config["optimizer"]["grad_method"] == "AD" else False,
+        bounds=ts_fitter.bounds,
+        options={"disp": True, "maxiter": 100},
+    )
+
+    best_loss = res["fun"]
+    best_weights = ts_fitter.unravel_pytree(res["x"])
+
+    return best_loss, best_weights
+
+
+def one_d_loop(config, all_data, sa, batch_indices, num_batches):
+    sample = {k: v[: config["optimizer"]["batch_size"]] for k, v in all_data.items()}
+    sample = {
+        "noise_e": config["other"]["PhysParams"]["noiseE"][: config["optimizer"]["batch_size"]],
+        "noise_i": config["other"]["PhysParams"]["noiseI"][: config["optimizer"]["batch_size"]],
+    } | sample
+    ts_fitter = TSFitter(config, sa, sample)
+
+    print("minimizing")
+    mlflow.set_tag("status", "minimizing")
     batch_indices = np.reshape(batch_indices, (-1, config["optimizer"]["batch_size"]))
+    all_weights = []
     overall_loss = 0.0
-
     previous_weights = None
     with trange(num_batches, unit="batch") as tbatch:
         for i_batch in tbatch:
@@ -161,150 +336,27 @@ def scipy_1d_loop(config, all_data, batch_indices, num_batches, best_weights, sa
                 "noise_e": config["other"]["PhysParams"]["noiseE"][inds],
                 "noise_i": config["other"]["PhysParams"]["noiseI"][inds],
             }
-            ts_fitter = TSFitter(config, sa, batch)
 
-            if previous_weights is None:
-                init_weights = ts_fitter.flattened_weights
+            if config["optimizer"]["method"] == "adam":  # Stochastic Gradient Descent
+                best_loss, best_weights = _1d_adam_loop_(config, ts_fitter, previous_weights, batch, tbatch)
+            else:
+                # not sure why this is needed but something needs to be reset, either the weights or the bounds
+                ts_fitter = TSFitter(config, sa, batch)
+                best_loss, best_weights = _1d_scipy_loop_(config, ts_fitter, previous_weights, batch, tbatch)
 
+            all_weights.append(best_weights)
+            mlflow.log_metrics({"batch loss": float(best_loss)}, step=i_batch)
+            overall_loss += best_loss
+
+            # ugly
             if "sequential" in config["optimizer"]:
                 if config["optimizer"]["sequential"]:
-                    if previous_weights is not None:
-                        init_weights = previous_weights
+                    if config["optimizer"]["method"] == "adam":
+                        previous_weights = best_weights
+                    else:
+                        previous_weights, _ = ravel_pytree(best_weights)
 
-            res = spopt.minimize(
-                ts_fitter.vg_loss if config["optimizer"]["grad_method"] == "AD" else ts_fitter.loss,
-                init_weights,
-                args=batch,
-                method=config["optimizer"]["method"],
-                jac=True if config["optimizer"]["grad_method"] == "AD" else False,
-                bounds=ts_fitter.bounds,
-                options={"disp": True, "maxiter": 100},
-            )
-            best_weights[i_batch] = ts_fitter.unravel_pytree(res["x"])
-            overall_loss += res["fun"]
-
-            if "sequential" in config["optimizer"]:
-                if config["optimizer"]["sequential"]:
-                    previous_weights = res.x
-
-        raw_weights = best_weights
-
-    return raw_weights, best_weights, overall_loss, ts_fitter
-
-
-def adam_loop(config, all_data, sa, batch_indices, num_batches):
-    if config["other"]["extraoptions"]["spectype"] == "angular_full":
-        config["optimizer"]["batch_size"] = 1
-        test_batch = {
-            "e_data": all_data["e_data"][config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :],
-            "e_amps": all_data["e_amps"][config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :],
-            "i_data": all_data["i_data"],
-            "i_amps": all_data["i_amps"],
-            "noise_e": config["other"]["PhysParams"]["noiseE"][
-                config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :
-            ],
-            "noise_i": config["other"]["PhysParams"]["noiseI"][
-                config["data"]["lineouts"]["start"] : config["data"]["lineouts"]["end"], :
-            ],
-        }
-        func_dict = get_loss_function(config, sa, test_batch)
-        jaxopt_kwargs = dict(
-            fun=func_dict["vg_func"], maxiter=config["optimizer"]["num_epochs"], value_and_grad=True, has_aux=True
-        )
-        opt = optax.adam(config["optimizer"]["learning_rate"])
-        solver = jaxopt.OptaxSolver(opt=opt, **jaxopt_kwargs)
-
-        weights = func_dict["init_weights"]
-        opt_state = solver.init_state(weights, batch=test_batch)
-
-        # start train loop
-        t1 = time.time()
-        print("minimizing")
-        mlflow.set_tag("status", "minimizing")
-
-        best_loss = 1e16
-        for i_epoch in (pbar := trange(config["optimizer"]["num_epochs"])):
-            if config["nn"]["use"]:
-                np.random.shuffle(batch_indices)
-                # tbatch.set_description(f"Epoch {i_epoch + 1}, Prev Epoch Loss {epoch_loss:.2e}")
-            epoch_loss = 0.0
-            weights, opt_state = solver.update(params=weights, state=opt_state, batch=test_batch)
-            epoch_loss += opt_state.value
-            pbar.set_description(f"Loss {epoch_loss:.2e}")
-
-            # epoch_loss /= num_batches
-            if epoch_loss < best_loss:
-                best_loss = epoch_loss
-                best_weights = weights
-
-            mlflow.log_metrics({"epoch loss": float(epoch_loss)}, step=i_epoch)
-
-    else:
-        test_batch = {k: v[config["optimizer"]["batch_size"]] for k, v in all_batches.items()}
-        func_dict = get_loss_function(config, sa, test_batch)
-        jaxopt_kwargs = dict(
-            fun=func_dict["vg_func"], maxiter=config["optimizer"]["num_epochs"], value_and_grad=True, has_aux=True
-        )
-        opt = optax.adam(config["optimizer"]["learning_rate"])
-        solver = jaxopt.OptaxSolver(opt=opt, **jaxopt_kwargs)
-
-        weights = func_dict["init_weights"]
-        opt_state = solver.init_state(weights, batch=test_batch)
-
-        # start train loop
-        t1 = time.time()
-        print("minimizing")
-        mlflow.set_tag("status", "minimizing")
-
-        epoch_loss = 1e19
-        best_loss = 1e16
-        for i_epoch in range(config["optimizer"]["num_epochs"]):
-            if config["nn"]["use"]:
-                np.random.shuffle(batch_indices)
-            batch_indices = np.reshape(batch_indices, (-1, config["optimizer"]["batch_size"]))
-            with trange(num_batches, unit="batch") as tbatch:
-                tbatch.set_description(f"Epoch {i_epoch + 1}, Prev Epoch Loss {epoch_loss:.2e}")
-                epoch_loss = 0.0
-                for i_batch in tbatch:
-                    inds = batch_indices[i_batch]
-                    batch = {
-                        "data": all_data["data"][inds],
-                        "amps": all_data["amps"][inds],
-                        "noise_e": config["other"]["PhysParams"]["noiseE"][inds],
-                        "noise_i": config["other"]["PhysParams"]["noiseI"][inds],
-                    }
-                    weights, opt_state = solver.update(params=weights, state=opt_state, batch=batch)
-                    epoch_loss += opt_state.value
-                    tbatch.set_postfix({"Prev Batch Loss": opt_state.value})
-
-                epoch_loss /= num_batches
-                if epoch_loss < best_loss:
-                    best_loss = epoch_loss
-                    best_weights = weights
-
-                mlflow.log_metrics({"epoch loss": float(epoch_loss)}, step=i_epoch)
-            batch_indices = batch_indices.flatten()
-
-    raw_weights = best_weights
-
-    overall_loss = epoch_loss
-
-    return raw_weights, best_weights, overall_loss, func_dict
-
-
-def scipy_loop(config, all_data, sa, batch_indices, num_batches):
-    best_weights = {}
-    all_weights = []
-    if config["other"]["extraoptions"]["spectype"] == "angular_full":
-        raw_weights, best_weights, overall_loss, ts_fitter = scipy_angular_loop(
-            config, all_data, best_weights, all_weights, sa
-        )
-    else:
-        raw_weights, best_weights, overall_loss, ts_fitter = scipy_1d_loop(
-            config, all_data, batch_indices, num_batches, best_weights, sa
-        )
-    mlflow.log_metrics({"overall loss": float(overall_loss)})
-    return raw_weights, best_weights, overall_loss, ts_fitter
+    return all_weights, overall_loss, ts_fitter
 
 
 def fit(config) -> Tuple[pd.DataFrame, float]:
@@ -333,7 +385,7 @@ def fit(config) -> Tuple[pd.DataFrame, float]:
     """
     t1 = time.time()
     mlflow.set_tag("status", "preprocessing")
-    config = validate_inputs(config)
+    config = _validate_inputs_(config)
 
     # prepare data
     all_data, sa, all_axes = prepare.prepare_data(config)
@@ -344,20 +396,16 @@ def fit(config) -> Tuple[pd.DataFrame, float]:
     # perform fit
     t1 = time.time()
     mlflow.set_tag("status", "minimizing")
-    if config["optimizer"]["method"] == "adam":  # Stochastic Gradient Descent
-        raw_weights, best_weights, overall_loss, func_dict = adam_loop(
-            config, all_data, sa, batch_indices, all_batches, num_batches
-        )
-    else:
-        raw_weights, best_weights, overall_loss, func_dict = scipy_loop(
-            config, all_data, sa, batch_indices, num_batches
-        )
 
+    if "angular" in config["other"]["extraoptions"]["spectype"]:
+        fitted_weights, overall_loss, ts_fitter = scipy_angular_loop(config, all_data, sa)
+    else:
+        fitted_weights, overall_loss, ts_fitter = one_d_loop(config, all_data, sa, batch_indices, num_batches)
+
+    mlflow.log_metrics({"overall loss": float(overall_loss)})
     mlflow.log_metrics({"fit_time": round(time.time() - t1, 2)})
     mlflow.set_tag("status", "postprocessing")
 
-    final_params = postprocess.postprocess(
-        config, batch_indices, all_data, all_axes, best_weights, func_dict, sa, raw_weights
-    )
+    final_params = postprocess.postprocess(config, batch_indices, all_data, all_axes, ts_fitter, sa, fitted_weights)
 
     return final_params, float(overall_loss)
