+++ inverse_thomson_scattering/model/loss_function.py
@@ -1,6 +1,5 @@
 import copy
 from typing import Dict
-from jax import config as jax_config
 
 import jax
 from jax import numpy as jnp
@@ -31,7 +30,6 @@ class TSFitter:
 
         self.spec_calc = SpectrumCalculator(cfg, sas, dummy_batch)
 
-        rng_key = jax.random.PRNGKey(42)
         self._loss_ = jit(self.__loss__)
         self._vg_func_ = jit(value_and_grad(self.__loss__, argnums=0, has_aux=True))
         self._h_func_ = jit(jax.hessian(self._loss_for_hess_fn_, argnums=0))
@@ -40,7 +38,10 @@ class TSFitter:
         ############
 
         if cfg["optimizer"]["method"] == "adam":
-            pass
+            lb, ub, init_weights = init_weights_and_bounds(cfg, num_slices=cfg["optimizer"]["batch_size"])
+            self.static_params = init_weights["inactive"]
+            self.pytree_weights = init_weights
+
         else:
             lb, ub, init_weights = init_weights_and_bounds(cfg, num_slices=cfg["optimizer"]["batch_size"])
             self.flattened_weights, self.unravel_pytree = ravel_pytree(init_weights["active"])
@@ -63,7 +64,7 @@ class TSFitter:
             else:
                 raise NotImplementedError
         else:
-            print("\n !!! Distribution function not fitted !!! Make sure this is what you thought you were running \n")
+            Warning("\n !!! Distribution function not fitted !!! Make sure this is what you thought you were running \n")
 
     def smooth(self, distribution):
         s = jnp.r_[
@@ -75,9 +76,12 @@ class TSFitter:
             self.smooth_window_len - 1 : -(self.smooth_window_len - 1)
         ]
 
-    def weights_to_params(self, these_params, static=True):
+    def weights_to_params(self, these_params, return_static_params=True):
         for param_name, param_config in self.cfg["parameters"].items():
             if param_config["active"]:
+                # if self.cfg["optimizer"]["method"] == "adam":
+                #     these_params[param_name] = 0.5 + 0.5 * jnp.tanh(these_params[param_name])
+
                 these_params[param_name] = (
                     these_params[param_name] * self.cfg["units"]["norms"][param_name]
                     + self.cfg["units"]["shifts"][param_name]
@@ -86,7 +90,7 @@ class TSFitter:
                     these_params["fe"] = jnp.log(self.smooth(jnp.exp(these_params["fe"][0]))[None, :])
 
             else:
-                if static:
+                if return_static_params:
                     these_params[param_name] = self.static_params[param_name]
 
         return these_params
@@ -105,19 +109,27 @@ class TSFitter:
         if self.cfg["other"]["extraoptions"]["fit_IAW"]:
             sqdev["ion"] = jnp.square(i_data - ThryI) / (jnp.abs(i_data) + 1e-1)
             sqdev["ion"] = jnp.where(
-                ((lamAxisI > self.cfg["data"]["fit_rng"]["iaw_min"])
-                & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_cf_min"]))
-                |((lamAxisI > self.cfg["data"]["fit_rng"]["iaw_cf_max"])
-                & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_max"])),
+                (
+                    (lamAxisI > self.cfg["data"]["fit_rng"]["iaw_min"])
+                    & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_cf_min"])
+                )
+                | (
+                    (lamAxisI > self.cfg["data"]["fit_rng"]["iaw_cf_max"])
+                    & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_max"])
+                ),
                 sqdev["ion"],
                 0.0,
             )
             loss += jnp.sum(sqdev["ion"], axis=1)
             used_points += jnp.sum(
-                ((lamAxisI > self.cfg["data"]["fit_rng"]["iaw_min"])
-                & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_cf_min"]))
-                |((lamAxisI > self.cfg["data"]["fit_rng"]["iaw_cf_max"])
-                & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_max"]))
+                (
+                    (lamAxisI > self.cfg["data"]["fit_rng"]["iaw_min"])
+                    & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_cf_min"])
+                )
+                | (
+                    (lamAxisI > self.cfg["data"]["fit_rng"]["iaw_cf_max"])
+                    & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_max"])
+                )
             )
 
         if self.cfg["other"]["extraoptions"]["fit_EPWb"]:
@@ -162,17 +174,19 @@ class TSFitter:
         e_data = batch["e_data"]
         if self.cfg["other"]["extraoptions"]["fit_IAW"]:
             _error_ = jnp.square(i_data - ThryI) / denom[0]
-            #print(jnp.shape(_error_))
-            #print(jnp.shape(lamAxisI))
             _error_ = jnp.where(
-                ((lamAxisI > self.cfg["data"]["fit_rng"]["iaw_min"])
-                & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_cf_min"]))
-                |((lamAxisI > self.cfg["data"]["fit_rng"]["iaw_cf_max"])
-                & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_max"])),
+                (
+                    (lamAxisI > self.cfg["data"]["fit_rng"]["iaw_min"])
+                    & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_cf_min"])
+                )
+                | (
+                    (lamAxisI > self.cfg["data"]["fit_rng"]["iaw_cf_max"])
+                    & (lamAxisI < self.cfg["data"]["fit_rng"]["iaw_max"])
+                ),
                 _error_,
                 0.0,
             )
-            
+
             i_error += reduce_func(_error_)
 
         if self.cfg["other"]["extraoptions"]["fit_EPWb"]:
@@ -224,7 +238,7 @@ class TSFitter:
         return fe_penalty
 
     def _loss_for_hess_fn_(self, params, batch):
-        params = {**params, **self.static_params}
+        params = params | self.static_params
         ThryE, ThryI, lamAxisE, lamAxisI = self.spec_calc(params, batch)
         i_error, e_error = self.calc_ei_error(
             batch,
@@ -254,9 +268,10 @@ class TSFitter:
         # other_losses = calc_other_losses(params)
         normed_batch = self.get_normed_batch(batch)
         normed_e_data = normed_batch["e_data"]
-        return self.cfg["data"][
-            "ion_loss_scale"
-        ] * i_error + e_error + density_loss + temperature_loss + momentum_loss, [ThryE, normed_e_data, params]
+        ion_error = self.cfg["data"]["ion_loss_scale"] * i_error
+
+        total_loss = ion_error + e_error + density_loss + temperature_loss + momentum_loss
+        return total_loss, [ThryE, normed_e_data, params]
 
     def get_normed_batch(self, batch):
         normed_batch = copy.deepcopy(batch)
